+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Upcoming
+* Renamed `Prevent Sleeping` to `Sleeping`
+  * You can now sleep during day, waking up at dusk
 * `Food Heal Multiplier` can now be configured even if Combat Test preset it enabled
 * `Crops Require Water` no longer affects crops that aren't on farmland (e.g. Farmer's Delight rice)